Terminado

+++ backend/e_commerce_app/models.py
@@ -44,8 +44,9 @@ class Producto(models.Model):
     tipoProducto = models.ForeignKey(TipoProducto, on_delete=models.CASCADE)
 
     def delete(self, *args, **kwargs):
-        if os.path.isfile(self.imagen.path):
-            os.remove(self.imagen.path)
+        if self.imagen != "":
+            if os.path.isfile(self.imagen.path):
+                os.remove(self.imagen.path)
         super(Producto, self).delete(*args, **kwargs)
 
     def __str__(self):
